requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
colorama==0.4.4
cryptography==37.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-browser-reload==1.3.0
django-ckeditor==6.4.0
django-etc==1.3.1
django-heroku==0.3.1
django-hitcount==1.3.5
django-imagefit==0.6.2
django-imagekit==2.0.4
django-js-asset==2.0.0
django-restframework==0.0.1
django-social-share==2.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
rave-python==1.2.15
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
